requirements.txt
django==5.0.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
mutagen==1.47.0
pydub==0.25.1
gunicorn==21.2.0
